% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelApplicationFunctions.R
\name{apply_mod_seasonal}
\alias{apply_mod_seasonal}
\title{Apply model to historical remotely sensed record}
\usage{
apply_mod_seasonal(histdata, model, season, threshold)
}
\arguments{
\item{histdata}{dataframe with reflectance values}

\item{model}{GLM (.Rdata)}

\item{season}{vector of months to include in the season}

\item{threshold}{numeric value at which to limit the allowed estimate)}
}
\description{
Apply GLM to remotely sensed record
}
