% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_vars.R
\name{create_model_vars}
\alias{create_model_vars}
\title{Perform Stepwise Regression}
\usage{
create_model_vars(filename, rowIndex = TRUE)
}
\arguments{
\item{filename}{CSV file with formatted surface reflectance data}

\item{rowIndex}{True or False, Indicates whether the first column in the formatted data file is a row index}
}
\value{
dataframe with variables used in model development or application
}
\description{
Use stepwise regression to parameterize initial model
}
