% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendFunctions.R
\name{doy_max_chl}
\alias{doy_max_chl}
\title{DOY for Max Chl}
\usage{
doy_max_chl(record, lake)
}
\arguments{
\item{record}{dataframe with estimated historical record of Chl-a levels}

\item{lake}{string, Name of Lake}
}
\description{
Calculates DOY for maximum chl-a for each station
}

