% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendFunctions.R
\name{extreme_trends}
\alias{extreme_trends}
\title{Extreme Trends}
\usage{
extreme_trends(record, lake, extreme = mean(record$FieldValue),
  plot = "station")
}
\arguments{
\item{record}{dataframe with estimated historical record of Chl-a levels}

\item{lake}{string, Name of Lake}

\item{extreme}{numerical value of the threshold above which are "extreme" levels}
}
\description{
Calculates trend and statistical significance for occurrence of extreme values
}
