% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateModel.R,
%   R/ModelParameterizationFunctions.R
\name{limitparams}
\alias{limitparams}
\alias{limitparams}
\title{Limit Parameters}
\usage{
limitparams(form, pval)

limitparams(form, pval)
}
\arguments{
\item{form}{Formula}

\item{pval}{P-value (confidence level)}

\item{form}{Formula}

\item{pval}{P-value (confidence level)}
}
\value{
formula with significant parameters

formula with significant parameters
}
\description{
Create formula limited to parameters above a confidence level

Create formula limited to parameters above a confidence level
}
