% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameterizationFunctions.R
\name{modresults}
\alias{modresults}
\title{Model Results}
\usage{
modresults(model, data, title)
}
\arguments{
\item{model}{GLM model}

\item{data}{data for calibration/evaluation}

\item{title}{character string, suffix for the title of the plot}
}
\description{
Evaluate Model Performance (R2 and RMSE)
}
