% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendFunctions.R
\name{monthly_trend}
\alias{monthly_trend}
\title{Explore linear trends}
\usage{
monthly_trend(record, lake)
}
\arguments{
\item{record}{dataframe with estimated historical record of chl-a levels}

\item{lake}{string, Name of Lake}
}
\description{
Calculates linear trend and significance on a monthly basis for each lake
}
