% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelApplicationFunctions.R
\name{plot_entire_record}
\alias{plot_entire_record}
\title{Plot record}
\usage{
plot_entire_record(histdata, obsdata, lake, labels = TRUE)
}
\arguments{
\item{histdata}{Dataframe with estimated values (FieldValue), dates (ImageDate), lake (Lake), lower and upper bounds (lower and upper)}

\item{obsdata}{Dataframe with Observed Data (Value, ImageDate)}

\item{lake}{string, Name of Lake}

\item{labels}{optional for plotting}
}
\description{
Plots estimated and observed data
}
