% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelApplicationFunctions.R
\name{plot_est_record.errors}
\alias{plot_est_record.errors}
\title{Plot record}
\usage{
plot_est_record.errors(histdata)
}
\arguments{
\item{histdata}{Dataframe with estimated values (FieldValue), dates (ImageDate), lake (Lake), lower and upper bounds (lower and upper)}
}
\description{
Plots estimated record with error bars
}

