% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendFunctions.R
\name{station_trend}
\alias{station_trend}
\title{Explore linear trends by station}
\usage{
station_trend(record)
}
\arguments{
\item{record}{dataframe with estimated historical record of chl-a levels}
}
\description{
Calculates linear trend and significance by station over the period of record
}
