% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrateModel.R,
%   R/ModelParameterizationFunctions.R
\name{stepChlModel}
\alias{stepChlModel}
\alias{stepChlModel}
\title{Perform Stepwise Regression}
\usage{
stepChlModel(caldata, timewindow, season, stepdirection, print.on = TRUE)

stepChlModel(caldata, timewindow, season, stepdirection, print.on = TRUE)
}
\arguments{
\item{caldata}{.Rda with field data and surface reflectance values}

\item{timewindow}{Number of days to allow for near coincidence}

\item{season}{Vector of months to include in model}

\item{stepdirection}{Direction for stepwise regression ("backward","both","forward")}

\item{caldata}{.Rda with field data and surface reflectance values}

\item{timewindow}{Number of days to allow for near coincidence}

\item{season}{Vector of months to include in model}

\item{stepdirection}{Direction for stepwise regression ("backward","both","forward")}
}
\value{
list with the stepwise model and the modeled values

list with the stepwise model and the modeled values
}
\description{
Use stepwise regression to parameterize initial model

Use stepwise regression to parameterize initial model
}
