% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParameterizationFunctions.R
\name{cvChlModel}
\alias{cvChlModel}
\title{Cross Validation}
\usage{
cvChlModel(caldf, k, model, gof)
}
\arguments{
\item{caldf}{dataframe limited to independent/dependent variables}

\item{k}{Number of folds (will not be used if there are fewer observations than folds)}

\item{form}{Formula for GLM}

\item{params}{Number of independent variables in the specified formula}
}
\value{
parameter estimates for specified model formula
}
\description{
Use k-fold or LOOCV to evaluate the R2, RMSE, and parameter estimates for a model
}
